Fix wipe function.
Due to the way the wipe function is used, any modern compiler would
optimize this call away.  To fix this we use the same macros as used
by Libgcrypt, GnuPG et. al.  The version from Libgcrypt has a
compatible license (LGPLv2.1) - we use an older version of that
because the newer fast_wipememory2 code is not required.

Patch submitted on behalf of Werner Koch <wk@gnupg.org>