build: Generate the export header
* src/CMakeLists.txt: Remove compile definition BUILDING_GPGMEPP. Remove
gpgmepp_export.h from gpgmepp_HEADERS. Use generate_export_header
to generate gpgmepp_export.h. Create a copy of the export header in a
gpgme++/ folder in the build folder. Install the generated header.
* src/gpgmepp_export.h: Remove.
* src/interfaces/{assuantransaction.h,dataprovider.h,statusconsumer.h}:
Include export header with gpgme++/gpgmepp_export.h.
--

This ensures that we have correct EXPORT and DEPRECATED macros for all
platforms/compilers.

During the build the interfaces headers cannot include the export
header anymore using a relative path because the export header is
generated in the build folder. Therefore we change the includes to
<gpgme++/gpgmepp_export.h> and make sure that those includes also work
for the build by creating a copy of the export header in a gpgme++/
folder in the build folder.

GnuPG-bug-id: 7262